Reorganized expositions header folder and its components
