+++ web/src/containers/expositions/header/fields/ViewOnlyPinnedCheckbox.tsx
