+++ web/src/containers/expositions/header/fields/old/SearchTextField.tsx
